requirements.txt
absl-py==2.0.0
aiofiles==22.1.0
aiosqlite==0.18.0
altair==5.1.2
amqp==2.6.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
bleach==6.0.0
blinker==1.6.3
cachetools==5.3.1
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.7
clx-sdk-xms==1.0.0
cmake==3.25.2
colorama==0.4.6
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-celery-beat==2.1.0
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==23.3
django-nested-admin==4.0.2
django-nested-inline==0.4.6
django-timezone-field==4.2.3
django-widget-tweaks==1.4.8
dlib==19.24.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.2
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
googletrans==3.0.0
grpcio==1.59.0
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
iso8601==1.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
keras==2.14.0
kiwisolver==1.4.5
kombu==4.6.11
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.7
mistune==2.0.5
ml-dtypes==0.2.0
MouseInfo==0.1.3
mss==9.0.1
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
npm==0.1.1
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-contrib-python-headless==4.7.0.68
opencv-python==4.8.1.78
opt-einsum==3.3.0
optional-django==0.1.0
outcome==1.3.0.post0
packaging==23.0
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
playsound==1.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-json-logger==2.0.4
python-monkey-business==1.0.0
pytweening==1.0.7
pytz==2020.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
react==4.3.0
redis==3.5.3
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.6.0
rsa==4.9
selenium==4.14.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post10
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
sqlparse==0.3.1
stack-data==0.6.2
streamlit==1.28.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.22.0
vine==1.3.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wordcloud==1.9.2
wrapt==1.14.1
wsproto==1.2.0
y-py==0.5.5
ypy-websocket==0.8.2
zipp==3.17.0